feat: remove keyGenerator and fix keyId format
The keyId is now fixed to be '<keyPrefix>:regular:<chatId>:<msgId>', where msgId
is the message_id of the sent message obtained from the return value of prev() in
the API transformer. This msgId is also the same as ctx.msgId when a callbackQuery
update arrives as a result of clicking on a menu message's callback button.

Users can no longer supply a custom keyGenerator function. The keyPrefix option
remains customizable.